+++ include/sf33rd/Source/Common/MemMan.h
@@ -6,13 +6,13 @@
 
 void mmSystemInitialize();
 void mmHeapInitialize(_MEMMAN_OBJ *mmobj, u8 *adrs, s32 size, s32 unit, s8 *format);
-u32 mmRoundUp(s32 unit, u32 num);
-u32 mmRoundOff(s32 unit, u32 num);
+uintptr_t mmRoundUp(s32 unit, uintptr_t num);
+uintptr_t mmRoundOff(s32 unit, uintptr_t num);
 void mmDebWriteTag(s8 * /* unused */);
-s32 mmGetRemainder(_MEMMAN_OBJ *mmobj);
-s32 mmGetRemainderMin(_MEMMAN_OBJ *mmobj);
-u8 *mmAlloc(_MEMMAN_OBJ *mmobj, s32 size, s32 flag);
-struct _MEMMAN_CELL *mmAllocSub(_MEMMAN_OBJ *mmobj, s32 size, s32 flag);
+ssize_t mmGetRemainder(_MEMMAN_OBJ *mmobj);
+ssize_t mmGetRemainderMin(_MEMMAN_OBJ *mmobj);
+u8 *mmAlloc(_MEMMAN_OBJ *mmobj, ssize_t size, s32 flag);
+struct _MEMMAN_CELL *mmAllocSub(_MEMMAN_OBJ *mmobj, ssize_t size, s32 flag);
 void mmFree(_MEMMAN_OBJ *mmobj, u8 *adrs);
 
 #endif
