+++ src/anniversary/sf33rd/AcrSDK/common/fbms.c
@@ -1,7 +1,7 @@
 #include "common.h"
 #include "structs.h"
 
-s32 fmsCalcSpace(FL_FMS *lp) {
+ptrdiff_t fmsCalcSpace(FL_FMS *lp) {
     return lp->frame[1] - lp->frame[0];
 }
 
@@ -9,8 +9,8 @@ s32 fmsInitialize(FL_FMS *lp, void *memory_ptr, s32 memsize, s32 memalign) {
     memsize = ~(memalign - 1) & (memsize + memalign - 1);
     lp->memoryblock = (u8 *)memory_ptr;
     lp->align = memalign;
-    lp->baseandcap[0] = (u8 *)((u32)lp->memoryblock + lp->align - 1 & ~(lp->align - 1));
-    lp->baseandcap[1] = (u8 *)(((u32)lp->memoryblock + memsize + lp->align - 1) & ~(lp->align - 1));
+    lp->baseandcap[0] = (u8 *)((uintptr_t)lp->memoryblock + lp->align - 1 & ~(lp->align - 1));
+    lp->baseandcap[1] = (u8 *)(((uintptr_t)lp->memoryblock + memsize + lp->align - 1) & ~(lp->align - 1));
     lp->frame[0] = lp->baseandcap[0];
     lp->frame[1] = lp->baseandcap[1];
     return 1;
@@ -18,7 +18,7 @@ s32 fmsInitialize(FL_FMS *lp, void *memory_ptr, s32 memsize, s32 memalign) {
 
 void *fmsAllocMemory(FL_FMS *lp, s32 bytes, s32 heapnum) {
     void *pMem;
-    bytes = ~(lp->align - 1) & ((u32)bytes + lp->align - 1);
+    bytes = ~(lp->align - 1) & ((uintptr_t)bytes + lp->align - 1);
 
     if (lp->frame[0] + bytes > lp->frame[1]) {
         return NULL;
